cenas

+++ Trabalho de PD Servidor/src/trabalho/de/pd/servidor/EliminarFicheiro.java
@@ -9,6 +9,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.ObjectOutputStream;
 import java.io.OutputStream;
 import java.net.Socket;
 import java.util.logging.Level;
@@ -18,22 +19,32 @@ import java.util.logging.Logger;
  *
  * @author ASUS
  */
-public class EliminarFicheiro extends Thread{
-    
-        
-    Servidor servidor=null;
-    Socket socketPedido=null;
-    String nomeFicheiro=null;
-    
-    public EliminarFicheiro(Servidor servidor,String nomeFicheiro) {
+public class EliminarFicheiro extends Thread {
+
+    Servidor servidor = null;
+    Socket socketPedido = null;
+    Pedido pedido=null;
+
+    public EliminarFicheiro(Servidor servidor, Pedido pedido) {
         this.servidor = servidor;
-        this.nomeFicheiro=nomeFicheiro;
+        this.pedido = pedido;
     }
-    
+
     @Override
     public void run() {
-        String dir=(servidor.getDiretoria());
-        File file=new File(dir);
-        file.delete();
+        if (servidor.isPrimario()) {
+            String dir = (servidor.getDiretoria());
+            File file = new File(dir);
+            file.delete();
+        }else{
+            ObjectOutputStream oos;
+            try {
+                oos = new ObjectOutputStream(servidor.getPrimarioSocketTCP().getOutputStream());
+                oos.writeObject(pedido);
+            } catch (IOException ex) {
+                Logger.getLogger(EliminarFicheiro.class.getName()).log(Level.SEVERE, null, ex);
+            }
+            
+        }
     }
 }
